requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pdfplumber>=0.10.0
openai>=1.0.0
python-multipart>=0.0.9
boto3>=1.34.0
google-auth>=2.25.0
google-api-python-client>=2.100.0
